Changes to support Node Local extents. (#725)
* make sure server main thread handles RM cleanup

* Remove consistency model config; 1.0 doc updates

Remove consistency model configuration code as this is an old
placeholder and UnifyFS will likely never claim POSIX compliance.

Update docs to prepare to 1.0 release.

- Update docs metadata
- Add limitation paragraph on directory support
- Remove mentions of the consistency model config option
- Minor adjustments for readability

* unifyfs-ls util (#736)

* Add new RPC to get GFIDs from the server

This is a first attempt at code to support a new RPC that will allow the
client to request a list of all the GFIDs that the server knows about.

This is still a work in progress with lots of TODO comments in the code!

* Style cleanups from checkpatch.sh

Minor clean-ups suggested by the checkpatch script.  Mostly things like
"int *foo" -> "int* foo".

* Minor cleanups to the get_gfids RPC code

Some minor cleanups: stop passing context structs and margo structs to
code that doesn't need them.  Clean up some error handling.  RPC enums
are supposed to be alphabetized.

* New API functions for getting file metadata

Add two new API functions: one to get the list of GFIDs from the server
(using the new get_gfids RPC) and one to get metadata for a specific
GFID (using the metaget RPC).

* Add new unifyfs-ls util

Add a new utility to do show file info similar to 'ls'.  The util is
still very preliminary and needs a number of refinements before it will
be ready for use by users.

* Fix style errors reported by checkpatch

* Unit tests for new API functions

Add unit tests for the new API functions: unifyfs_get_gfid_list() and
unifyfs_get_server_file_meta().

Note: test of the file size (part of the data returned by
unifyfs_get_server_file_meta()) fails.  Still checking to see if this is
a problem in the unit test or in the actual code.

* checkpatch.sh cleanups

* Improvements to unifyfs-ls

Add a few command line options and generally make the program less hacky
and more usable for non-developers.

* Fix unit test for unifyfs_get_server_file_meta()

Call unifyfs_syc() after completing the writes for each file so that the
server knows the correct size to report.

* Comment out 2 checks in unit test

Because other unit tests don't always clean up after themselves, we
can't be sure how many gfids will be returned by
unifyfs_get_gfid_list().  So, comment out the test for unexpected gfids
and replace the test for the exact number of gfids with one that only
check for *at least* a certain number.

* Minor checkpatch fixes

Fixed a few errors regarding "int *foo" instead of "int* foo"

* Re-arrange error handling code in invoke_client_get_gfids_rpc()

The error handling inside invoke_client_get_gfids_rpc() was not correct
for all cases.

* Clean up style problems flagged by checkpatch

Clean up a variety of problems flagged by the checkpatch script - mostly
lines >80 characters.

* Install unifyfs-ls in the bin dir

Updates Makefile.am to install the unifyfs-ls util in the bin dir
instead of libexec.

* Add documentation for unifyfs-ls

Update run.rst to add user-level documentation for the unifyfs-ls tool.

* Update docs/run.rst

Remove full path from example command line


* Fix improper error handling

Fix 2 cases of incorrect error handling:  one where it was possible to
call free() on an unitialized variable and one where memory didn't get
freed when it should have been.


* Add CXXFLAGS to unifyfs-ls

Add -std=c++11 CXXFLAGS to the unifyfs-ls target in order to
build on GCC 4.9.3.

* Changes to support Node Local extents.

- The client has a new configuration for node_local_extents
- On first read of a laminated file if client.node_local_extents is
  enabled we fetch extents from server and update the local extents
- This allows reads to be executed locally
- on server we get our extents using unifyfs_invoke_find_extents_rpc
- compiler sets value to null by default
- Reformatted the code to pass the format check test on Github
- In case of distributed I need to make sure the iteration
  is correct.
- Added a client level configuration for enabling node local
  extents for laminated files.
- fixed the type definition to use * explicitly
  except for the margo macros which need typedef.
- removed node-local extents from init-fini.c
- fixed using rc instead of ret which is used
  for return of functions
- add a null check.
- added NULL check for accessing metadata
- renamed unifyfs_filemeta_t.needs_sync
  -> unifyfs_filemeta_t.needs_write_sync
- renamed unifyfs_filemeta_t.needs_extents_sync
  -> unifyfs_filemeta_t.needs_reads_sync
- Newer glibc malloc requires parameters to be size_t
- the request now uses a array of unifyfs_extent_t
- the response uses an array of unifyfs_client_index_t
- we still need the list as we calculate extents across
  all request and response

In these cases we do only one rpc per request group and fetch extents
for all files which are laminated.

+++ client/src/client_read.h
@@ -111,6 +111,6 @@ void update_read_req_coverage(read_req_t* req,
 /* process a set of client read requests */
 int process_gfid_reads(unifyfs_client* client,
                        read_req_t* in_reqs,
-                       int in_count);
+                       size_t in_count);
 
 #endif // UNIFYFS_CLIENT_READ_H
